Moves noPorxy merge func to a separate pkg
