+++ pkg/controller/proxyconfig/status.go
@@ -3,16 +3,11 @@ package proxyconfig
 import (
 	"context"
 	"fmt"
-	"net/url"
-	"strconv"
-	"strings"
-
-	"github.com/ghodss/yaml"
 
 	configv1 "github.com/openshift/api/config/v1"
+	"github.com/openshift/cluster-network-operator/pkg/util/proxyconfig"
 
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/util/sets"
 )
 
 // syncProxyStatus computes the current status of proxy and
@@ -22,11 +17,11 @@ func (r *ReconcileProxyConfig) syncProxyStatus(proxy *configv1.Proxy, infra *con
 	var noProxy string
 	updated := proxy.DeepCopy()
 
-	if isSpecNoProxySet(&proxy.Spec) {
+	if isSpecNoProxySet(&proxy.Spec) || isSpecHTTPProxySet(&proxy.Spec) || isSpecHTTPSProxySet(&proxy.Spec) {
 		if proxy.Spec.NoProxy == noProxyWildcard {
 			noProxy = proxy.Spec.NoProxy
 		} else {
-			noProxy, err = mergeUserSystemNoProxy(proxy, infra, network, cluster)
+			noProxy, err = proxyconfig.MergeUserSystemNoProxy(proxy, infra, network, cluster)
 			if err != nil {
 				return fmt.Errorf("failed to merge user/system noProxy settings: %v", err)
 			}
@@ -44,72 +39,6 @@ func (r *ReconcileProxyConfig) syncProxyStatus(proxy *configv1.Proxy, infra *con
 	}
 
 	return nil
-}
-
-// mergeUserSystemNoProxy merges user-supplied noProxy settings from proxy
-// with cluster-wide noProxy settings, returning a merged, comma-separated
-// string of noProxy settings.
-func mergeUserSystemNoProxy(proxy *configv1.Proxy, infra *configv1.Infrastructure, network *configv1.Network, cluster *corev1.ConfigMap) (string, error) {
-	apiServerURL, err := url.Parse(infra.Status.APIServerURL)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse API server URL")
-	}
-
-	internalAPIServer, err := url.Parse(infra.Status.APIServerInternalURL)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse API server internal URL")
-	}
-
-	set := sets.NewString(
-		"127.0.0.1",
-		"localhost",
-		network.Status.ServiceNetwork[0],
-		apiServerURL.Hostname(),
-		internalAPIServer.Hostname(),
-	)
-
-	// TODO: This will be flexible when master machine management is more dynamic.
-	type installConfig struct {
-		ControlPlane struct {
-			Replicas string `json:"replicas"`
-		} `json:"controlPlane"`
-		Networking struct {
-			MachineCIDR string `json:"machineCIDR"`
-		} `json:"networking"`
-	}
-	var ic installConfig
-	data, ok := cluster.Data["install-config"]
-	if !ok {
-		return "", fmt.Errorf("missing install-config in configmap")
-	}
-	if err := yaml.Unmarshal([]byte(data), &ic); err != nil {
-		return "", fmt.Errorf("invalid install-config: %v\njson:\n%s", err, data)
-	}
-
-	switch infra.Status.PlatformStatus.Type {
-	case configv1.AWSPlatformType, configv1.GCPPlatformType, configv1.AzurePlatformType, configv1.OpenStackPlatformType:
-		set.Insert("169.254.169.254", ic.Networking.MachineCIDR)
-	}
-
-	replicas, err := strconv.Atoi(ic.ControlPlane.Replicas)
-	if err != nil {
-		return "", fmt.Errorf("failed to parse install config replicas: %v", err)
-	}
-
-	for i := int64(0); i < int64(replicas); i++ {
-		etcdHost := fmt.Sprintf("etcd-%d.%s", i, infra.Status.EtcdDiscoveryDomain)
-		set.Insert(etcdHost)
-	}
-
-	for _, clusterNetwork := range network.Status.ClusterNetwork {
-		set.Insert(clusterNetwork.CIDR)
-	}
-
-	for _, userValue := range strings.Split(proxy.Spec.NoProxy, ",") {
-		set.Insert(userValue)
-	}
-
-	return strings.Join(set.List(), ","), nil
 }
 
 // proxyStatusesEqual compares two ProxyStatus values. Returns true if the
